# swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Library Management API",
        "description": "Library Management System API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Development Server"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/books/Update": {
            "put": {
                "tags": [
                    "book-controller"
                ],
                "operationId": "updateBook",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookUpdateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseBookResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/book-requests/Update": {
            "put": {
                "tags": [
                    "book-request-controller"
                ],
                "operationId": "updateBookRequest",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookRequestUpdateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseBookRequestResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/books/Create": {
            "post": {
                "tags": [
                    "book-controller"
                ],
                "operationId": "createBook",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookCreationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseBookResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/book-requests/Create": {
            "post": {
                "tags": [
                    "book-request-controller"
                ],
                "operationId": "createBookRequest",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookRequestCreationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseBookRequestResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/books/GetById": {
            "get": {
                "tags": [
                    "book-controller"
                ],
                "operationId": "getBookById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseBookResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/GetAll": {
            "get": {
                "tags": [
                    "book-controller"
                ],
                "operationId": "getAllBooks",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "publisherId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "authorId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "categoryId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "skipCount",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "default": 0
                        }
                    },
                    {
                        "name": "maxResultCount",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseBaseGetAllResponseBookResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/book-requests/GetById": {
            "get": {
                "tags": [
                    "book-request-controller"
                ],
                "operationId": "getBookRequestById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseBookRequestResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/book-requests/GetAll": {
            "get": {
                "tags": [
                    "book-request-controller"
                ],
                "operationId": "getAllBookLoans",
                "parameters": [
                    {
                        "name": "bookTitle",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ACCEPTED",
                                "DENIED",
                                "PENDING"
                            ]
                        }
                    },
                    {
                        "name": "userId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "username",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "BORROWING",
                                "RETURNING",
                                "EXTENDING"
                            ]
                        }
                    },
                    {
                        "name": "skipCount",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "default": 0
                        }
                    },
                    {
                        "name": "maxResultCount",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseBaseGetAllResponseBookRequestResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/book-loans/GetAll": {
            "get": {
                "tags": [
                    "book-loan-controller"
                ],
                "operationId": "getAllBookLoans_1",
                "parameters": [
                    {
                        "name": "bookTitle",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "REJECTED",
                                "REQUEST_BORROWING",
                                "BORROWED",
                                "REQUEST_RETURNING",
                                "RETURNED",
                                "NONRETURNABLE"
                            ]
                        }
                    },
                    {
                        "name": "userId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "skipCount",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "default": 0
                        }
                    },
                    {
                        "name": "maxResultCount",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseBaseGetAllResponseBookLoanResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/books/Delete": {
            "delete": {
                "tags": [
                    "book-controller"
                ],
                "operationId": "deleteBook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseString"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/book-requests/Delete": {
            "delete": {
                "tags": [
                    "book-request-controller"
                ],
                "operationId": "deleteBookRequest",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseString"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiResponseString": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "result": {
                        "type": "string"
                    },
                    "error": {
                        "type": "object"
                    }
                }
            },
...
            "BookUpdateRequest": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string",
                        "minLength": 1
                    },
                    "price": {
                        "type": "number",
                        "format": "double"
                    },
                    "description": {
                        "type": "string"
                    },
                    "publisherId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "authorIds": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "uniqueItems": true
                    },
                    "categoryIds": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "uniqueItems": true
                    }
                },
                "required": [
                    "authorIds",
                    "categoryIds",
                    "price",
                    "publisherId"
                ]
            },
            "ApiResponseBookResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "result": {
                        "$ref": "#/components/schemas/BookResponse"
                    },
                    "error": {
                        "type": "object"
                    }
                }
            },
            "Author": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    }
                }
            },
            "BookResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number",
                        "format": "double"
                    },
                    "description": {
                        "type": "string"
                    },
                    "publisherId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "numberOfCopiesAvailable": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "publisherName": {
                        "type": "string"
                    },
                    "authors": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Author"
                        },
                        "uniqueItems": true
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        },
                        "uniqueItems": true
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "BookRequestUpdateRequest": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "ACCEPTED",
                            "DENIED",
                            "PENDING"
                        ]
                    }
                },
                "required": [
                    "id"
                ]
            },
            "ApiResponseBookRequestResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "result": {
                        "$ref": "#/components/schemas/BookRequestResponse"
                    },
                    "error": {
                        "type": "object"
                    }
                }
            },
            "BookRequestResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "bookCopyId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "bookTitle": {
                        "type": "string"
                    },
                    "userId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "username": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "BORROWING",
                            "RETURNING",
                            "EXTENDING"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "ACCEPTED",
                            "DENIED",
                            "PENDING"
                        ]
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "BookCreationRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "minLength": 1
                    },
                    "price": {
                        "type": "number",
                        "format": "double"
                    },
                    "description": {
                        "type": "string"
                    },
                    "publisherId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "authorIds": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "uniqueItems": true
                    },
                    "categoryIds": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "uniqueItems": true
                    }
                },
                "required": [
                    "authorIds",
                    "categoryIds",
                    "price",
                    "publisherId"
                ]
            },
            "BookRequestCreationRequest": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "ACCEPTED",
                            "DENIED",
                            "PENDING"
                        ]
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "BORROWING",
                            "RETURNING",
                            "EXTENDING"
                        ]
                    },
                    "bookLoanId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
            "ApiResponseBaseGetAllResponseBookResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "result": {
                        "$ref": "#/components/schemas/BaseGetAllResponseBookResponse"
                    },
                    "error": {
                        "type": "object"
                    }
                }
            },
            "BaseGetAllResponseBookResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BookResponse"
                        }
                    },
                    "totalRecords": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "ApiResponseBaseGetAllResponseBookRequestResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "result": {
                        "$ref": "#/components/schemas/BaseGetAllResponseBookRequestResponse"
                    },
                    "error": {
                        "type": "object"
                    }
                }
            },
            "BaseGetAllResponseBookRequestResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BookRequestResponse"
                        }
                    },
                    "totalRecords": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "ApiResponseBaseGetAllResponseBookLoanResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "result": {
                        "$ref": "#/components/schemas/BaseGetAllResponseBookLoanResponse"
                    },
                    "error": {
                        "type": "object"
                    }
                }
            },
            "BaseGetAllResponseBookLoanResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BookLoanResponse"
                        }
                    },
                    "totalRecords": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}